Todo Class 외부 파일로 분기
- Class 분리로 이름이 중복되어 헷갈림
  - todos => data로 변경
